include/streamlit/requirements.txt
streamlit==1.33.0
weaviate-client==4.5.5
openai==1.19.0